[REF][runbot] Add validation of module duplicate

+++ runbot/runbot.py
@@ -572,7 +572,8 @@ class runbot_build(osv.osv):
             # v6 rename bin -> openerp
             if os.path.isdir(build.path('bin/addons')):
                 shutil.move(build.path('bin'), build.server())
-
+            
+            
             # fallback for addons-only community/project branches
             if not os.path.isfile(build.server('__init__.py')):
                 # Use modules to test previously configured in the repository
@@ -590,10 +591,14 @@ class runbot_build(osv.osv):
                 # Finally move all addons to openerp/addons
                 for module in glob.glob(build.path('*/__openerp__.py')):
                     shutil.move(os.path.dirname(module), build.path('openerp/addons'))
-
+            
             # move all addons to server addons path
             for i in glob.glob(build.path('addons/*')):
-                shutil.move(i, build.server('addons'))
+                if not os.path.exists( build.server('addons', os.path.basename( i ) ) ):
+                    shutil.move(i, build.server('addons'))
+                else:
+                    build._log('Building environment', 'You have next module duplicate into your branches "%s"'%( os.path.basename( i ) ) )
+                
 
     def pg_dropdb(self, cr, uid, dbname):
         pid_col = 'pid' if cr._cnx.server_version >= 90200 else 'procpid'
@@ -654,6 +659,7 @@ class runbot_build(osv.osv):
         return cmd, modules
 
     def spawn(self, cmd, lock_path, log_path, cpu_limit=None, shell=False, showstderr=False):
+        print "cmd", cmd#TODO: Add to log runbot
         def preexec_fn():
             os.setsid()
             if cpu_limit:
